requirements.txt
detectron2==0.5
gpiozero==1.5.1
opencv_python==4.5.1.48
requests==2.21.0
adafruit_circuitpython_ads1x15==2.2.8
Flask==1.0.2
pandas==1.2.4
torchvision==0.8.0a0+45f960c
seaborn==0.11.1
smbus==1.1.post2
tqdm==4.61.0
adafruit_circuitpython_bh1750==1.0.4
Adafruit_Blinka==6.6.2
scipy==1.6.3
pafy==0.5.5
numpy==1.16.2
matplotlib==3.4.2
torch==1.7.0a0
adafruit_ads1x15==1.0.2
albumentations==1.0.3
board==1.0
onnxruntime==1.8.1
paho_mqtt==1.5.1
Pillow==8.3.1
protobuf==3.17.3
PyYAML==5.4.1
thop==0.0.31-2005241907
train==0.0.5
